requirements.txt
# PlantVillage Rose Edition - Dependencies
# ML Frameworks
ultralytics>=8.0.0
sahi>=0.11.0
torch>=2.0.0
torchvision>=0.15.0
tensorflow>=2.13.0

# Web Interface
streamlit>=1.28.0

# Image Processing
opencv-python>=4.8.0
Pillow>=9.5.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# Dataset Access
datasets>=2.14.0
kagglehub>=0.1.0
roboflow>=1.1.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
requests>=2.31.0

# Testing
pytest>=7.4.0
